chat-service/requirements.txt
Django>=5.2
channels>=4.0
channels-redis>=4.1
daphne>=4.0
psycopg2-binary>=2.9
PyJWT==2.8.0
python-dotenv>=1.0
djangorestframework
django-cors-headers